Refactored code for all metrics in one file and middleware in a file
