+++ server/lomas_server/routes/utils.py
@@ -1,5 +1,3 @@
-import json
-import logging
 import random
 import time
 from collections.abc import AsyncGenerator
@@ -18,67 +16,11 @@ from lomas_core.models.requests import (
     QueryModel,
 )
 from lomas_core.models.responses import CostResponse, QueryResponse
-from opentelemetry.trace import get_tracer
-from starlette.middleware.base import BaseHTTPMiddleware
 
 from lomas_server.data_connector.factory import data_connector_factory
 from lomas_server.dp_queries.dp_libraries.factory import querier_factory
 from lomas_server.dp_queries.dummy_dataset import get_dummy_dataset_for_query
 from lomas_server.utils.config import get_config
-
-
-class LoggingAndTracingMiddleware(BaseHTTPMiddleware):
-    """
-    Middleware for logging and tracing incoming HTTP requests.
-
-    This middleware logs the incoming requests, including the user name
-    the route being accessed, and any query parameters.
-    Additionally, it creates a trace span to trace the user’s request and
-    adds attributes to the span related to the user name and query parameters.
-    """
-
-    async def dispatch(self, request: Request, call_next):
-        """
-        Handles the request and performs logging and tracing.
-
-        Logs the user name, the route and the query parameters.
-        Creates a trace span to monitor the request and adds relevant attributes.
-
-        Args:
-            request (Request): The incoming request object.
-            call_next (Callable): A function that, when called, passes the request
-                                  to the next middleware or request handler.
-
-        Returns:
-            Response: The HTTP response generated by calling `call_next(request)`.
-        """
-        user_name = request.headers.get("user-name")
-        route = request.url.path
-
-        try:
-            query_params = await request.json()
-        except (json.JSONDecodeError, ValueError):
-            query_params = {}
-        for param, value in query_params.items():
-            if value is None:
-                query_params[param] = ""
-            if isinstance(value, dict):
-                query_params[param] = json.dumps(value)
-
-        logging.info(
-            f"User '{user_name}' is making a request to route '{route}' "
-            + f"with query params: {query_params}."
-        )
-
-        tracer = get_tracer(__name__)
-        with tracer.start_as_current_span("user_request_span") as span:
-            span.set_attribute("user_name", user_name)
-            for param, value in query_params.items():
-                span.set_attribute(f"query_param.{param}", value)
-
-            response = await call_next(request)
-
-        return response
 
 
 def timing_protection(func):
